+++ CHANGES.rst
@@ -4,6 +4,10 @@ Change log
 2.2 (unreleased)
 ----------------
 
+- Add ``--with-free-threaded-python`` option to enable CI testing with the
+  free-threaded (nogil) build of CPython.
+  (`#388 <https://github.com/zopefoundation/meta/issues/388>`_)
+
 - Remove special treatment for macOS test runners on GitHub
 
 - Print out a warning if a ``setuptools`` install time dependency is found